extract list to separate component
